fix: refactored file picker code in seperate component
