+++ app/client/src/pages/common/datasourceAuth/index.tsx
@@ -17,7 +17,6 @@ import {
   setDatasourceViewMode,
   createDatasourceFromForm,
   toggleSaveActionFlag,
-  filePickerCallbackAction,
 } from "actions/datasourceActions";
 import AnalyticsUtil from "utils/AnalyticsUtil";
 import { getCurrentApplicationId } from "selectors/editorSelectors";
@@ -25,7 +24,6 @@ import { useParams, useLocation } from "react-router";
 import type { ExplorerURLParams } from "@appsmith/pages/Editor/Explorer/helpers";
 import type { AppState } from "@appsmith/reducers";
 import type { Datasource } from "entities/Datasource";
-import { FilePickerActionStatus } from "entities/Datasource";
 import { AuthType, AuthenticationStatus } from "entities/Datasource";
 import {
   CONFIRM_CONTEXT_DELETING,
@@ -46,7 +44,6 @@ import {
   hasDeleteDatasourcePermission,
   hasManageDatasourcePermission,
 } from "@appsmith/utils/permissionHelpers";
-import { removeHTMLBodyOverlay } from "pages/utils";
 
 interface Props {
   datasource: Datasource;
@@ -60,8 +57,6 @@ interface Props {
   triggerSave?: boolean;
   isFormDirty?: boolean;
   datasourceDeleteTrigger: () => void;
-  gsheetToken?: string;
-  gsheetProjectID?: string;
 }
 
 export type DatasourceFormButtonTypes = Record<string, string[]>;
@@ -122,8 +117,6 @@ function DatasourceAuth({
   shouldDisplayAuthMessage = true,
   triggerSave,
   isFormDirty,
-  gsheetToken,
-  gsheetProjectID,
 }: Props) {
   const authType =
     formData && "authType" in formData
@@ -157,17 +150,8 @@ function DatasourceAuth({
   const pageId = (pageIdQuery || pageIdProp) as string;
   const [confirmDelete, setConfirmDelete] = useState(false);
 
-  const [scriptLoadedFlag] = useState<boolean>(
-    (window as any).googleAPIsLoaded,
-  );
-  const [pickerInitiated, setPickerInitiated] = useState<boolean>(false);
-  const [pickerVisible, setPickerVisible] = useState<boolean>(false);
   const dsName = datasource?.name;
   const orgId = datasource?.workspaceId;
-
-  // objects gapi and google are set, when google apis script is loaded
-  const gapi: any = (window as any).gapi;
-  const google: any = (window as any).google;
 
   useEffect(() => {
     if (confirmDelete) {
@@ -305,76 +289,6 @@ function DatasourceAuth({
     }
   };
 
-  useEffect(() => {
-    // Since we need to display file picker on blank page, as soon as file picker is visible
-    // Add overlay on the file picker background
-    if (pickerVisible) {
-      const element: HTMLElement | null =
-        document.querySelector(".picker-dialog-bg");
-      if (!!element) {
-        element.style.opacity = "1";
-      }
-    }
-  }, [pickerVisible]);
-
-  useEffect(() => {
-    // This loads the picker object in gapi script
-    if (!!gsheetToken && !!gapi && !!gsheetProjectID) {
-      gapi.load("client:picker", async () => {
-        await gapi.client.load(
-          "https://www.googleapis.com/discovery/v1/apis/drive/v3/rest",
-        );
-        setPickerInitiated(true);
-      });
-    }
-  }, [scriptLoadedFlag, gsheetToken, gsheetProjectID]);
-
-  useEffect(() => {
-    if (
-      !!gsheetToken &&
-      scriptLoadedFlag &&
-      pickerInitiated &&
-      !!google &&
-      !!gsheetProjectID
-    ) {
-      createPicker(gsheetToken, gsheetProjectID);
-    }
-  }, [gsheetToken, scriptLoadedFlag, pickerInitiated, gsheetProjectID]);
-
-  const createPicker = async (accessToken: string, projectID: string) => {
-    const view = new google.picker.View(google.picker.ViewId.SPREADSHEETS);
-    view.setMimeTypes("application/vnd.google-apps.spreadsheet");
-    const picker = new google.picker.PickerBuilder()
-      .enableFeature(google.picker.Feature.NAV_HIDDEN)
-      .enableFeature(google.picker.Feature.MULTISELECT_ENABLED)
-      .setAppId(projectID)
-      .setOAuthToken(accessToken)
-      .addView(view)
-      .setCallback(pickerCallback)
-      .build();
-    picker.setVisible(true);
-    setPickerVisible(true);
-  };
-
-  const pickerCallback = async (data: any) => {
-    // Remove document body overlay as soon as file picker is loaded
-    if (data.action === FilePickerActionStatus.LOADED) {
-      removeHTMLBodyOverlay();
-    } else if (
-      data.action === FilePickerActionStatus.CANCEL ||
-      data.action === FilePickerActionStatus.PICKED
-    ) {
-      setPickerVisible(false);
-      const fileIds = data?.docs?.map((element: any) => element.id) || [];
-      dispatch(
-        filePickerCallbackAction({
-          action: data.action,
-          datasourceId: datasourceId,
-          fileIds: fileIds,
-        }),
-      );
-    }
-  };
   const createMode = datasourceId === TEMP_DATASOURCE_ID;
 
   const datasourceButtonsComponentMap = (buttonType: string): JSX.Element => {
